Create all repositoris and some updates in models and dto's

+++ Propelo/DTO/ApartmentDTO.cs
@@ -8,7 +8,7 @@ namespace Propelo.DTO
         public string Name { get; set; }
         public type Type { get; set; }
         public int Floor { get; set; }
-        public int Surface { get; set; }
+        public double Surface { get; set; }
         public string Description { get; set; }
     }
 }
